requirements.txt
google-generativeai
geocoder